refactor/test_utils: move TestError to test_utils
This makes TestError accessible in integration tests. This is useful because
TestError implements ErrorCode and the test code doesn't have to go through the
boilerplate of creating an error type.
